requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --strip-extras
#
pyyaml==6.0.2
    # via watchdog
watchdog==6.0.0
    # via -r requirements.in
